requirements.txt
anthropic>=0.40.0
feedparser>=6.0.0
requests>=2.31.0
beautifulsoup4>=4.12.0
pyyaml>=6.0
apscheduler>=3.10.0

# dev
pytest>=8.0.0
ruff>=0.8.0
